[JENKINS-69789] Address review feedback for password complexity rule (#27038)
* refactor: Address post-merge review feedback from daniel-beck

- Move password complexity classes from hudson.security to jenkins.security
- Fix special character regex: use \p{Punct} instead of [^a-zA-Z0-9]
  so spaces and Unicode letters are no longer treated as special chars
- Document special character examples on the config UI
- Add violations list to PasswordComplexityException for better UX
- Remove redundant checkMethod="post" (already the default)
- Move message keys to jenkins/security/Messages.properties
- Update copyright year to 2026


* fix: Remove unused NonePasswordComplexityRule import


* refactor: Add @Restricted(NoExternalUse.class) to built-in implementations

Plugins only need to extend PasswordComplexityRule — there is no need
to reference BasicPasswordComplexityRule or NonePasswordComplexityRule
directly.


* docs: Update @since to acknowledge package move from hudson.security

Per daniel-beck's feedback: document that these classes were originally
added in 2.572 and moved from hudson.security to jenkins.security in
this release, so plugin authors are aware of the breaking change.


---------
