+++ core/src/main/resources/hudson/security/HudsonPrivateSecurityRealm/Details/config.jelly
@@ -25,9 +25,9 @@ THE SOFTWARE.
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
     <f:entry title="${%Password}:">
-      <input class="jenkins-input validated" checkMethod="post" checkUrl="${descriptor.getCheckMethod('password').toStemUrl()}" checkDependsOn="" name="user.password" type="password" value="${instance.protectedPassword}" />
+      <input class="jenkins-input validated" checkUrl="${descriptor.getCheckMethod('password').toStemUrl()}" checkDependsOn="" name="user.password" type="password" value="${instance.protectedPassword}" />
     </f:entry>
     <f:entry title="${%Confirm Password}:">
-      <input class="jenkins-input validated" checkMethod="post" checkUrl="${descriptor.getCheckMethod('password').toStemUrl()}" checkDependsOn="" name="user.password2" type="password" value="${instance.protectedPassword}" />
+      <input class="jenkins-input validated" checkUrl="${descriptor.getCheckMethod('password').toStemUrl()}" checkDependsOn="" name="user.password2" type="password" value="${instance.protectedPassword}" />
     </f:entry>
 </j:jelly>
