+++ test/src/test/java/hudson/security/HudsonPrivateSecurityRealmTest.java
@@ -55,6 +55,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.logging.Level;
 import jenkins.security.ApiTokenProperty;
+import jenkins.security.BasicPasswordComplexityRule;
 import jenkins.security.SecurityListener;
 import jenkins.security.apitoken.ApiTokenPropertyConfiguration;
 import jenkins.security.seed.UserSeedProperty;
